Refactored to use BigDecimal for amount and price

+++ model/src/main/java/com/gtc/model/tradinggateway/api/dto/data/OrderDto.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Created by Valentyn Berezin on 16.01.18.
@@ -13,8 +14,8 @@ import java.io.Serializable;
 public class OrderDto implements Serializable {
 
     private String orderId;
-    private double size;
-    private double price;
+    private BigDecimal size;
+    private BigDecimal price;
     private OrderStatus status;
     private String statusString;
 }
