+++ src/main/java/com/gtc/tradinggateway/service/CreateOrder.java
@@ -3,6 +3,8 @@ package com.gtc.tradinggateway.service;
 import com.gtc.tradinggateway.meta.TradingCurrency;
 import com.gtc.tradinggateway.service.dto.OrderCreatedDto;
 
+import java.math.BigDecimal;
+
 /**
  * Typically WSS based.
  */
@@ -11,5 +13,5 @@ public interface CreateOrder extends ClientNamed {
     /**
      * @return created order id.
      */
-    OrderCreatedDto create(TradingCurrency from, TradingCurrency to, double amount, double price);
+    OrderCreatedDto create(TradingCurrency from, TradingCurrency to, BigDecimal amount, BigDecimal price);
 }
