+++ src/main/java/com/gtc/tradinggateway/service/hitbtc/dto/HitbtcBalanceItemDto.java
@@ -2,6 +2,8 @@ package com.gtc.tradinggateway.service.hitbtc.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * Created by mikro on 13.02.2018.
  */
@@ -9,5 +11,5 @@ import lombok.Data;
 public class HitbtcBalanceItemDto {
 
     private String currency;
-    private double available;
+    private BigDecimal available;
 }
